# src/webapi/cnt_sp_webapi/3 UseCases/SP.WebApi.UseCases/Handlers/Category/Queries/ListCategories/Responses/ListCategoriesResponse.cs
using SP.WebApi.UseCases.Handlers.Category.Dto;

namespace SP.WebApi.UseCases.Handlers.Category.Queries.ListCategories.Responses;

/// <summary>
/// Ответ со списком категорий.
/// </summary>
public sealed record ListCategoriesResponse(IReadOnlyList<CategoryDto> Items);